feat: implement java-block-mix

+++ de.tu-bs.cs.isf.cbc.util/src/de/tu_bs/cs/isf/cbc/util/HashTable.java
@@ -53,7 +53,8 @@ public class HashTable {
 				}
 				String fileName = line[0];
 				String fileHash = line[1];
-				if (fileName.equals(filename)) {
+//				System.out.println("checking filename " + fileName + " and " + filename);
+				if (fileName.equals(filename) || (fileName.contains(".java") && (fileName.split("\\.")[0] + ".class").equals(filename))) {
 					fileReader.close();
 					return true;
 				}
@@ -88,7 +89,9 @@ public class HashTable {
 	            if (!fileExistsInOverview(location, file.getName()) && 
 	            		!file.getName().equals(OVERVIEW_FILENAME) &&
 	            		!file.getName().equals(OVERVIEW_TMP_FILENAME) &&
-	            		!file.getName().equals("helper.key")
+	            		!file.getName().equals("helper.key") &&
+	            		!file.getName().equals("GlobalJavaVariables.java") &&
+	            		!file.getName().equals("GlobalJavaVariables.class")
 	            		) {
 		    	    System.out.println("delete" + file.getName());
 	            	file.delete();
